+++ CHANGELOG.md
@@ -29,6 +29,8 @@
     - [#527](https://github.com/wix-incubator/rich-content/pull/527) remove onTab (moved to keyBindingFn/handleKeyCommand)
   - `pubsub`
     - [#528](https://github.com/wix-incubator/rich-content/pull/528) rename visibleBlock to focusedBlock
+  - `button`
+    - [#517](https://github.com/wix-incubator/rich-content/pull/517) plugin button refactor
   
 </details>
 <hr/>